+++ Text/URI/Parser/ByteString.hs
@@ -22,9 +22,11 @@ where
 
 import Control.Monad
 import Control.Monad.Catch (MonadThrow (..))
+import Control.Monad.State.Strict
 import Data.ByteString (ByteString)
 import Data.Char
 import Data.List (intercalate)
+import Data.List.NonEmpty (NonEmpty (..))
 import Data.Maybe (isJust, catMaybes, maybeToList)
 import Data.Text (Text)
 import Data.Void
@@ -145,16 +147,26 @@ pUserInfo = try $ do
   return UserInfo {..}
 {-# INLINE pUserInfo #-}
 
-pPath :: MonadParsec e ByteString m => Bool -> m (Bool, [RText 'PathPiece])
+pPath :: MonadParsec e ByteString m
+  => Bool
+  -> m (Bool, Maybe (Bool, NonEmpty (RText 'PathPiece)))
 pPath hasAuth = do
   doubleSlash <- lookAhead (option False (True <$ string "//"))
   when (doubleSlash && not hasAuth) $
     (unexpected . Tokens . NE.fromList) [47,47]
   absPath <- option False (True <$ char 47)
-  path <- flip sepBy (char 47) . label "path piece" $
-    many pchar
-  pieces <- mapM (liftR mkPathPiece) (filter (not . null) path)
-  return (absPath, pieces)
+  (rawPieces, trailingSlash) <- flip runStateT False $
+    flip sepBy (char 47) . label "path piece" $ do
+      x <- many pchar
+      put (null x)
+      return x
+  pieces <- mapM (liftR mkPathPiece) (filter (not . null) rawPieces)
+  return
+    ( absPath
+    , case NE.nonEmpty pieces of
+        Nothing -> Nothing
+        Just ps -> Just (trailingSlash, ps)
+    )
 {-# INLINE pPath #-}
 
 pQuery :: MonadParsec e ByteString m => m [QueryParam]